backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
TensorFlow==2.14.0
Keras==2.14.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
gym==0.26.2
scipy==1.11.2
SQLAlchemy==2.0.19